port rpm cataloger to new generic cataloger pattern
