+++ syft/pkg/cataloger/cataloger.go
@@ -41,7 +41,7 @@ func ImageCatalogers(cfg Config) []pkg.Cataloger {
 		javascript.NewJavascriptPackageCataloger(),
 		javascript.NewNodeBinaryCataloger(),
 		deb.NewDpkgdbCataloger(),
-		rpm.NewRpmdbCataloger(),
+		rpm.NewRpmDBCataloger(),
 		java.NewJavaCataloger(cfg.Java()),
 		apkdb.NewApkdbCataloger(),
 		golang.NewGoModuleBinaryCataloger(),
@@ -61,7 +61,7 @@ func DirectoryCatalogers(cfg Config) []pkg.Cataloger {
 		javascript.NewJavascriptLockCataloger(),
 		javascript.NewNodeBinaryCataloger(),
 		deb.NewDpkgdbCataloger(),
-		rpm.NewRpmdbCataloger(),
+		rpm.NewRpmDBCataloger(),
 		rpm.NewFileCataloger(),
 		java.NewJavaCataloger(cfg.Java()),
 		java.NewJavaPomCataloger(),
@@ -90,7 +90,7 @@ func AllCatalogers(cfg Config) []pkg.Cataloger {
 		javascript.NewJavascriptPackageCataloger(),
 		javascript.NewNodeBinaryCataloger(),
 		deb.NewDpkgdbCataloger(),
-		rpm.NewRpmdbCataloger(),
+		rpm.NewRpmDBCataloger(),
 		rpm.NewFileCataloger(),
 		java.NewJavaCataloger(cfg.Java()),
 		java.NewJavaPomCataloger(),
